Add support for detailed monitoring of specific paths
If a whitelist of paths is specified via the -monitored_paths flag, the
exporter will record status counts at the granularity of {status, path,
method}. Paths are matched verbatim, rather than prefix, to the first
non-path sub-expression in the request (query params, fragment, etc.).

TODO(swfrench): Support prefix matching with '*' at the end of the
supplied path or the like. Alternatively, we can support a regexp.